Lwpmlse 377 l1 api64 single so (#1995)
* new ILP64 API for dot, dot_ex scal, scal_ex
* test infrastructure updates for new API testing both fortran and C 64bit forms
* single .so design but cmake code updates for potential future splitting of shared library

[ROCm/rocblas commit: 53c155babac346c51b684f7fda574e064bce5f93]

+++ clients/gtest/general_gtest.cpp
@@ -319,7 +319,7 @@ namespace
         // just want to ensure 64 bit math
 
         uint32_t block  = 8;
-        int64_t  stride = 1LL << 33; // int32 overflow
+        int64_t  stride = int64_t(1) << 33; // int32 overflow
         int64_t  offset = stride + 2;
 
         T        val(42);
